Remove #ifdefs for readability
